requirements.txt
imap-tools
python-dotenv
requests
pdfplumber
pytesseract
pdf2image